Use proper naming convention for Software Freedom Conservancy
While many people use "SFC" to refer to Software Freedom Conservancy,
the organization itself uses "Conservancy".  It's just "Conservancy"
and not "The ... Conservancy".

I don't find a good reference for this usage but here's a change
to the Conservancy web site:
https://k.sfconservancy.org/website/changeset/12780197e5df85b433ab54c03871a98523a0db88

Note that http://sfconservancy.org/ doesn't use "SFC" anywhere.