Avoid ambiguity when refering to foundations
Use "Linux Foundation" insted of just "Linux" since "Linux" has too
many different uses that it's not clear it refers to the Linux
Foundation (for example, there's also kernel.org which might be
"Linux").

To be consistent, use the full name for Eclipse Foundation and
Apache Software Foundation.  The other abbreviations and short names
are common (Conservancy, SPI).